feat: add calculator tab for anix services
