+++ package.json
@@ -17,7 +17,9 @@
     "prop-types": "^15.8.1",
     "lucide-react": "^0.436.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "jspdf": "^2.5.1",
+    "react-router-dom": "^6.27.0"
   },
   "devDependencies": {
     "@tailwindcss/forms": "^0.5.9",
