+++ src/data/calculatorConfig.js
@@ -1,0 +1,228 @@
+export const productTypes = [
+  {
+    id: 'single',
+    label: 'Один обучающий ролик (1–5 мин)',
+    description: 'Глубокий обучающий ролик для внутренних команд или клиентов.',
+    basePrice: 240000,
+    pricePerMinute: 92000,
+    baseTimeline: 18,
+    timelinePerMinute: 0.8,
+    baseTeamHours: 120,
+    teamHoursPerMinute: 20,
+    baseTeamLoad: 46,
+    expectedImpact: 720000,
+    defaultDurationSeconds: 60,
+    defaultCreativeCount: 1,
+    baseSavings: { time: 10, budget: 6, risk: 8, hours: 18, money: 120000 },
+  },
+  {
+    id: 'multi',
+    label: 'Несколько креативов (5–15 сек)',
+    description: 'Пачка рекламных креативов для перфоманса и социальных сетей.',
+    basePrice: 110000,
+    pricePerMinute: 180000,
+    baseTimeline: 12,
+    timelinePerMinute: 0.4,
+    baseTeamHours: 70,
+    teamHoursPerMinute: 12,
+    baseTeamLoad: 38,
+    expectedImpact: 540000,
+    defaultDurationSeconds: 12,
+    defaultCreativeCount: 3,
+    supportsCreatives: true,
+    baseSavings: { time: 8, budget: 7, risk: 6, hours: 14, money: 95000 },
+  },
+  {
+    id: 'academy',
+    label: 'Обучение под ключ / пакет роликов',
+    description: 'Системный курс или серия модулей с контролем знаний.',
+    basePrice: 480000,
+    pricePerMinute: 78000,
+    baseTimeline: 32,
+    timelinePerMinute: 1.6,
+    baseTeamHours: 220,
+    teamHoursPerMinute: 32,
+    baseTeamLoad: 58,
+    expectedImpact: 980000,
+    defaultDurationSeconds: 180,
+    defaultCreativeCount: 4,
+    supportsCreatives: true,
+    baseSavings: { time: 14, budget: 11, risk: 10, hours: 32, money: 180000 },
+  },
+];
+
+export const speedOptions = [
+  { id: 'normal', label: 'Обычные сроки', priceMultiplier: 1, timeMultiplier: 1 },
+  { id: 'fast', label: 'Ускоренный (+20% к цене)', priceMultiplier: 1.2, timeMultiplier: 0.82 },
+  { id: 'rush', label: 'Срочный (+40% к цене)', priceMultiplier: 1.4, timeMultiplier: 0.68 },
+];
+
+export const formatOptions = [
+  {
+    id: 'fullhd',
+    label: 'Full HD (16:9)',
+    description: 'Основной мастер-файл в горизонтальном формате.',
+    price: 16000,
+    teamHours: 6,
+    defaultSelected: true,
+    timeline: 0,
+    teamLoad: 2,
+  },
+  {
+    id: 'vertical',
+    label: 'Вертикальный (9:16)',
+    description: 'Вертикальные адаптации для Reels, Shorts и Stories.',
+    price: 12000,
+    teamHours: 5,
+    perCreative: true,
+    timeline: 1,
+    teamLoad: 2,
+  },
+  {
+    id: '4k',
+    label: '4K',
+    description: 'Высокое разрешение для премиальных экранов и архивов.',
+    price: 24000,
+    teamHours: 6,
+    timeline: 1,
+    teamLoad: 2,
+  },
+  {
+    id: 'social',
+    label: 'Форматы для соцсетей и пересылок',
+    description: 'Пакет квадратных/кадреных версий и лёгких превью.',
+    price: 18000,
+    teamHours: 6,
+    perCreative: true,
+    timeline: 2,
+    teamLoad: 3,
+  },
+];
+
+export const serviceOptions = [
+  {
+    id: 'marketing',
+    label: '🎯 Маркетинговый анализ и исследование воронки',
+    description: 'Глубокий аудит точек контакта, гипотез роста и болей аудитории.',
+    price: 65000,
+    teamLoad: 6,
+    teamHours: 18,
+    timeline: 4,
+    clientSavings: { time: 15, budget: 6, risk: 15, hours: 14, money: 90000 },
+    roiBoost: 12,
+  },
+  {
+    id: 'utp',
+    label: '💡 Проработка УТП и сценария',
+    description: 'Формируем ключевой месседж и структуру будущего ролика.',
+    price: 74000,
+    teamLoad: 7,
+    teamHours: 22,
+    timeline: 3,
+    clientSavings: { time: 12, budget: 8, risk: 10, hours: 12, money: 80000 },
+    roiBoost: 10,
+    defaultSelected: true,
+  },
+  {
+    id: 'script',
+    label: '✍️ Разработка ТЗ и сценарного драфта',
+    description: 'Детализированное ТЗ, режиссёрский план и интерактивные правки.',
+    price: 82000,
+    teamLoad: 8,
+    teamHours: 28,
+    timeline: 5,
+    clientSavings: { time: 11, budget: 7, risk: 9, hours: 10, money: 70000 },
+    roiBoost: 9,
+    defaultSelected: true,
+  },
+  {
+    id: 'concept',
+    label: '🧩 Разработка концепта и визуального стиля',
+    description: 'Moodboard, раскадровка, цветовые и типографические решения.',
+    price: 91000,
+    pricePerMinute: 16000,
+    teamLoad: 9,
+    teamHours: 34,
+    hoursPerMinute: 6,
+    timeline: 6,
+    clientSavings: { time: 10, budget: 9, risk: 12, hours: 14, money: 90000 },
+    roiBoost: 11,
+    defaultSelected: true,
+  },
+  {
+    id: 'mascot',
+    label: '🦉 Создание маскота / персонажей',
+    description: 'Разработка героев и guideline по применению в коммуникациях.',
+    price: 68000,
+    teamLoad: 5,
+    teamHours: 20,
+    timeline: 4,
+    clientSavings: { time: 6, budget: 5, risk: 8, hours: 8, money: 60000 },
+    roiBoost: 7,
+  },
+  {
+    id: 'animation',
+    label: '🎨 Анимация и монтаж',
+    description: 'Полный цикл производства, композитинг и финальная сборка.',
+    price: 98000,
+    pricePerMinute: 22000,
+    teamLoad: 12,
+    teamHours: 60,
+    hoursPerMinute: 18,
+    timeline: 7,
+    clientSavings: { time: 9, budget: 6, risk: 10, hours: 16, money: 85000 },
+    roiBoost: 14,
+    defaultSelected: true,
+    perCreative: true,
+  },
+  {
+    id: 'voiceover',
+    label: '🎙️ Озвучка',
+    description: 'Профессиональная запись или нейросеть — выберите формат.',
+    teamLoad: 4,
+    teamHours: 10,
+    timeline: 2,
+    clientSavings: { time: 6, budget: 4, risk: 5, hours: 6, money: 30000 },
+    roiBoost: 6,
+    defaultSelected: true,
+    voiceOptions: [
+      { id: 'female', label: 'Женский голос', price: 36000 },
+      { id: 'male', label: 'Мужской голос', price: 36000 },
+      { id: 'neural', label: 'Нейросеть', price: 18000 },
+    ],
+    defaultVoice: 'female',
+  },
+  {
+    id: 'focus',
+    label: '🧪 Тест на фокус-группе',
+    description: 'Качественная проверка креатива на целевой аудитории.',
+    price: 42000,
+    teamLoad: 4,
+    teamHours: 14,
+    timeline: 3,
+    clientSavings: { time: 7, budget: 6, risk: 14, hours: 10, money: 65000 },
+    roiBoost: 8,
+  },
+  {
+    id: 'analytics',
+    label: '📊 Мониторинг метрик и UTM-аналитика',
+    description: 'Настройка дашбордов, событий и контроль эффективности.',
+    price: 36000,
+    teamLoad: 3,
+    teamHours: 16,
+    timeline: 2,
+    clientSavings: { time: 8, budget: 9, risk: 11, hours: 12, money: 72000 },
+    roiBoost: 9,
+  },
+  {
+    id: 'distribution',
+    label: '🧭 Рекомендации по дистрибуции и стратегии продвижения',
+    description: 'План запуска, медиамикс и чек-листы интеграций.',
+    price: 48000,
+    teamLoad: 4,
+    teamHours: 18,
+    timeline: 3,
+    clientSavings: { time: 9, budget: 10, risk: 12, hours: 12, money: 78000 },
+    roiBoost: 10,
+  },
+];
